move logical scalar function type-coercion logic from physical-expr to expr
