+++ datafusion/physical-expr/src/expressions/negative.rs
@@ -28,10 +28,9 @@ use arrow::{
     record_batch::RecordBatch,
 };
 
-use crate::coercion_rule::binary_rule::is_signed_numeric;
 use crate::PhysicalExpr;
 use datafusion_common::{DataFusionError, Result};
-use datafusion_expr::ColumnarValue;
+use datafusion_expr::{binary_rule::is_signed_numeric, ColumnarValue};
 
 /// Invoke a compute kernel on array(s)
 macro_rules! compute_op {
